install graph and apply to the application - gathering article from friend feature is completed

+++ api/src/main/java/com/timeline/api/repository/UserRepository.java
@@ -4,12 +4,12 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.timeline.api.entity.UserEntity;
 
-public interface UserRepository extends CrudRepository<UserEntity, Integer>{
+public interface UserRepository extends CrudRepository<UserEntity, Long>{
 
 	boolean existsByEmail(String email);
 	 		
 	UserEntity findByEmail(String email);
 	
-	UserEntity findById(int id);
+	UserEntity findById(long id);
 
 }
